src/functions/requirements.txt
azure-functions==1.21.3
httpx>=0.27.2
pydantic>=2.8.2
pydantic-settings>=2.4.0
python-dotenv>=1.0.1
sqlalchemy[asyncio]>=2.0.32
asyncpg>=0.29.0
azure-storage-queue>=12.9.0
openai>=1.52.0
anthropic>=0.39.0
google-generativeai>=0.7.2

